save episode as float so extra episodes that are meant to be inbetween 2 regular episodes are ordered correctly;

+++ src/episode.cpp
@@ -98,7 +98,7 @@ void Episode::setWatched(bool value) {
     emit watchedChanged(oldValue, value);
 }
 
-int Episode::getEpisodeNumber() const {
+float Episode::getEpisodeNumber() const {
     return this->episodeNumber;
 }
